% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-sem_cfa_growth.R
\name{brcfa}
\alias{brcfa}
\title{Fit confirmatory factor analysis model using bias-reducing methods}
\usage{
brcfa(
  model,
  data,
  estimator = "ML",
  estimator.args = list(rbm = "iRBM", plugin_penalty = pen_huber),
  information = c("expected", "observed", "first.order"),
  ...
)
}
\arguments{
\item{model}{A description of the user-specified model. Typically, the model
is described using the lavaan model syntax. See \link{model.syntax} for more
information. Alternatively, a parameter table (eg. the output of the
\code{\link[=lavaanify]{lavaanify()}} function) is also accepted.}

\item{data}{An optional data frame containing the observed variables used in
the model. If some variables are declared as ordered factors, lavaan will
treat them as ordinal variables.}

\item{estimator}{The estimator to use. \strong{Currently only "ML" is supported.}}

\item{estimator.args}{A list containing RBM arguments. Possible arguments are
\itemize{
\item \code{rbm}: The type of RBM method to use. One of \code{FALSE}, \code{"eRBM"}, or \code{"iRBM"}.
\item \code{plugin_penalty}: The type of penalty to use. One of \code{NULL}, \code{"pen_ridge"}, or \code{"pen_ridge_bound"}.
}}

\item{information}{The type of information matrix to use for both the bias
reduction method and the calculation of standard errors. One of "expected",
"observed", or "first.order".

However, when "first.order" is specified, then this is only for the
standard error calculations. The bias reduction methods will revert to
using the expected information.}

\item{...}{Additional arguments to pass to the \link{lavaan} function.}
}
\value{
An object of class \code{brlavaan} which is a subclass of the
\linkS4class{lavaan} class.
}
\description{
Fit confirmatory factor analysis model using bias-reducing methods
}
