% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-brlavaan.R
\name{brgrowth}
\alias{brgrowth}
\title{Fit growth curve models using empirical bias-reducing methods}
\usage{
brgrowth(
  model,
  data,
  estimator = "ML",
  estimator.args = list(rbm = "implicit", plugin_pen = NULL),
  information = "observed",
  ...
)
}
\arguments{
\item{model}{A description of the user-specified model. Typically, the model
is described using the lavaan model syntax. See \link[lavaan:model.syntax]{lavaan::model.syntax} for more
information. Alternatively, a parameter table (eg. the output of the
\code{\link[lavaan:model.syntax]{lavaan::lavaanify()}} function) is also accepted.}

\item{data}{An optional data frame containing the observed variables used in
the model. If some variables are declared as ordered factors, lavaan will
treat them as ordinal variables.}

\item{estimator}{The estimator to use. \strong{Currently only "ML" is supported.}}

\item{estimator.args}{A list containing RBM arguments. Possible arguments are
\itemize{
\item \code{rbm}: The type of RBM method to use. One of \code{"none"}, \code{"explicit"}, or
\code{"implicit"} (although, short forms are accepted, e.g. \code{"exp"}, \code{"iRBM"},
etc.)
\item \code{plugin_pen}: The type of penalty to use. One of \code{NULL}, \code{"pen_ridge"},
or \code{"pen_ridge_bound"}.
\item \code{info_pen} The type of information matrix to use for the penalty term.
\item \code{info_bias} The type of information matrix to use for the bias term of
the explicit reduced bias method.
}}

\item{information}{The type of information matrix to use for calculation of
standard errors. Defaults to \code{"observed"}, although \code{"expected"} and
\code{"first.order"} is also permitted.}

\item{...}{Additional arguments to pass to the \link[lavaan:lavaan]{lavaan::lavaan} function.}
}
\value{
An object of class \code{brlavaan} which is a subclass of the
\link[lavaan:lavaan-class]{lavaan::lavaan} class.
}
\description{
Fit growth curve models using empirical bias-reducing methods
}
