% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brlavaan-package.R
\docType{package}
\name{brlavaan-package}
\alias{brlavaan-package}
\title{brlavaan: Bias-Reduced Latent Variable Analysis}
\description{
Apply empirical bias reduced methods to fit a variety of latent variable models, including confirmatory factor analysis, structural equation modelling, and latent growth curve models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haziqj/sem-bias}
  \item \url{https://haziqj.ml/sem-bias/}
  \item Report bugs at \url{https://github.com/haziqj/sem-bias/issues}
}

}
\author{
\strong{Maintainer}: Haziq Jamil \email{haziq.jamil@gmail.com} (\href{https://orcid.org/0000-0003-3298-1010}{ORCID})

Other contributors:
\itemize{
  \item Yves Rosseel (\href{https://orcid.org/0000-0002-4129-4477}{ORCID}) [contributor]
  \item Oliver Kemp [contributor]
  \item Ioannis Kosmidis (\href{https://orcid.org/0000-0003-1556-0302}{ORCID}) [contributor]
}

}
\keyword{internal}
