% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-brlavaan.R
\name{brlavaan}
\alias{brlavaan}
\title{Fit a latent variable model using bias-reducing methods}
\usage{
brlavaan(
  model,
  data,
  estimator = "ML",
  estimator.args = list(rbm = "implicit", plugin_pen = NULL),
  information = "observed",
  lavfun = "sem",
  ...
)
}
\arguments{
\item{model}{A description of the user-specified model. Typically, the model
is described using the lavaan model syntax. See \link{model.syntax} for more
information. Alternatively, a parameter table (eg. the output of the
\code{\link[=lavaanify]{lavaanify()}} function) is also accepted.}

\item{data}{An optional data frame containing the observed variables used in
the model. If some variables are declared as ordered factors, lavaan will
treat them as ordinal variables.}

\item{estimator}{The estimator to use. \strong{Currently only "ML" is supported.}}

\item{estimator.args}{A list containing RBM arguments. Possible arguments are
\itemize{
\item \code{rbm}: The type of RBM method to use. One of \code{"none"}, \code{"explicit"}, or
\code{"implicit"} (although, short forms are accepted, e.g. \code{"exp"}, \code{"iRBM"},
etc.)
\item \code{plugin_pen}: The type of penalty to use. One of \code{NULL}, \code{"pen_ridge"},
or \code{"pen_ridge_bound"}.
\item \code{info_pen} The type of information matrix to use for the penalty term.
\item \code{info_bias} The type of information matrix to use for the bias term of
the explicit reduced bias method.
}}

\item{information}{The type of information matrix to use for calculation of
standard errors. Defaults to \code{"observed"}, although \code{"expected"} and
\code{"first.order"} is also permitted.}

\item{lavfun}{The lavaan function to use. Default is "sem".}

\item{...}{Additional arguments to pass to the \link{lavaan} function.}
}
\value{
An object of class \code{brlavaan} which is a subclass of the
\linkS4class{lavaan} class.
}
\description{
Fit a latent variable model using bias-reducing methods
}
\details{
The \code{pen_ridge()} function applies a ridge regression style penalty $f(x) =
|| x ||^2$ that shrinks the parameters to zero. The \code{pen_ridge_bound()}
function applies a penalty that shrinks the parameters to the bounds of the
parameter space. The bounds are calculated by \code{{lavaan}} -- see the paper for
more details.

The \code{info_pen} and \code{info_bias} arguments default to \code{"observed"}. It is not
recommended to change this to say \code{"expected"}, especially for small sample
sizes as the bias-reducing properties of the estimators are not guaranteed.
}
