% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41-manual_twofac.R
\name{fit_twofac}
\alias{fit_twofac}
\title{Fit two factor SEM using lavaan syntax manually}
\usage{
fit_twofac(model, data, rbm = c("none", "explicit", "implicit"), start = NULL)
}
\arguments{
\item{model}{A description of the user-specified model. Typically, the model
is described using the lavaan model syntax. See \link[lavaan:model.syntax]{lavaan::model.syntax} for more
information. Alternatively, a parameter table (eg. the output of the
\code{\link[lavaan:model.syntax]{lavaan::lavaanify()}} function) is also accepted.}

\item{data}{An optional data frame containing the observed variables used in
the model. If some variables are declared as ordered factors, lavaan will
treat them as ordinal variables.}

\item{rbm}{The type of RBM method to use. One of \code{"none"}, \code{"explicit"}, or
\code{"implicit"} (although, short forms are accepted, e.g. \code{"exp"}, \verb{"iRBM},
etc.)}

\item{start}{A numeric vector of starting values}
}
\value{
A list with the following components:
\item{coefficients}{The estimated coefficients}
\item{stderr}{The standard errors of the estimated coefficients}
\item{bias}{The bias of the estimated coefficients}
\item{timing}{The elapsed time}
\item{converged}{Whether the optimization converged}
\item{optim}{The optimization object}
}
\description{
Fit two factor SEM using lavaan syntax manually
}
