% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-gen_data.R
\name{growth-curve}
\alias{growth-curve}
\alias{txt_mod_growth_pop}
\alias{txt_mod_growth}
\title{Model specification for growth curve models}
\usage{
txt_mod_growth_pop(rel)

txt_mod_growth(rel)
}
\arguments{
\item{rel}{Reliability of the growth curve model. Either 0.8 or 0.5.}
}
\value{
A character string with the model specification, used for either
fitting lavaan models or simulating data.
}
\description{
Model specification for growth curve models
}
\examples{
txt_mod_growth_pop(0.8)
txt_mod_growth(0.5)
}
