% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-plugins.R
\name{plugin-penalties}
\alias{plugin-penalties}
\alias{pen_ridge}
\alias{pen_ridge_bound}
\alias{pen_huber}
\title{Functions for plugin penalty}
\usage{
pen_ridge(x, target = rep(0, length(x)), call = FALSE, ...)

pen_ridge_bound(x, lb, ub, call = FALSE, ...)

pen_huber(x, lb, ub, thres = 1, call = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{target}{The target values for components of \code{x} in the ridge penalty.
Default is a vector of zeros.}

\item{call}{Logical. If \code{TRUE}, return the name of the penalty function.}

\item{...}{Additional arguments which may be called. See Details.}

\item{lb}{The lower bounds for components of \code{x}.}

\item{ub}{The upper bounds for components of \code{x}.}

\item{thres}{The threshold for the Huber penalty.}
}
\value{
A single numeric value, which is the penalty value.
}
\description{
Functions for plugin penalty
}
