% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utilities.R
\name{predicates}
\alias{predicates}
\alias{is_ML}
\alias{is_eRBM}
\alias{is_iRBM}
\alias{is_iRBMp}
\title{Predicates for \code{brlavaan} objects}
\usage{
is_ML(x)

is_eRBM(x)

is_iRBM(x)

is_iRBMp(x, quietly = FALSE)
}
\arguments{
\item{x}{An object of class \code{brlavaan}, or an output from \code{\link[=fit_sem]{fit_sem()}}.}

\item{quietly}{If \code{TRUE}, suppresses messages.}
}
\value{
\code{TRUE} or \code{FALSE}, or if not suppressed, a message indicating which
plugin penalty was used.
}
\description{
Predicates for \code{brlavaan} objects
}
