% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-sim_functions.R
\name{sim_fun}
\alias{sim_fun}
\title{Conduct bias reduction simulation study}
\usage{
sim_fun(
  dist = "Normal",
  model = "twofac",
  rel = 0.8,
  n = 15,
  nsimu = 1,
  lavsim = FALSE,
  whichsims = c("ML", "eRBM", "iRBM"),
  info_pen = "observed",
  info_bias = "observed",
  info_se = "observed",
  keep_going = FALSE,
  data_scale = 1,
  seeds = NULL
)
}
\arguments{
\item{dist}{Distribution of the manifest variables. Either "Normal",
"Kurtosis" or "Non-normal".}

\item{model}{The model to simulate. Either \code{growth} or \code{twofac}.}

\item{rel}{Reliability of the growth curve model. Either 0.8 or 0.5.}

\item{n}{Sample size}

\item{nsimu}{The number of replications to run.}

\item{lavsim}{Logical. If TRUE, simulate data using lavaan's \link{simulateData}
function.}

\item{whichsims}{A character vector with the estimators to use. Possible
values are \code{ML}, \code{eRBM}, and \code{iRBM} for now}

\item{info_pen}{Should be \code{"observed"}}

\item{info_bias}{Should be \code{"observed"}}

\item{info_se}{Should be \code{"observed"}}

\item{keep_going}{If \code{TRUE}, the simulation will continue until the desired
\code{nsimu} runs are obtained.}

\item{data_scale}{A scaling factor for the data. Default is \code{1}. Usefor for
the growth curve simulations that this is set to \code{1/10}.}

\item{seeds}{A vector of seeds to use for the simulation study.}
}
\value{
A list with two elements. The first element is a data frame with the
results of the simulation study. The second element is a list with the
errors that occurred during the simulation study.
}
\description{
Conduct bias reduction simulation study
}
\details{
Suggest to enable parallel processing by calling
\code{future::plan("multisession", workers = future::availableCores() - 1)} before
running this function.

In our paper, we report on the results of the following simulation study. For
each of the two models a) growth curve model and b) two-factor model, we
generate a sample size of \code{n} observations from a distribution \code{dist} with
reliability \code{rel}. We then fit the model using the maximum likelihood (\code{ML})
estimator, the explicit RBM estimator (\code{eRBM}), the implicit RBM estimator
(\code{iRBM}), and the implicit RBM estimator with plug-in penalties (\code{iRBMp_x}).
We repeat this process \code{nsimu} times and report the results in terms of bias,
standard error, and computation time.
}
