% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-gen_data.R
\name{true-values}
\alias{true-values}
\alias{truth_growth}
\alias{truth_twofac}
\alias{truth}
\title{Retrieve the true values of the models used for simulation}
\usage{
truth_growth(rel, scale = 1)

truth_twofac(rel, meanstructure = FALSE, scale = 1)

truth(x)
}
\arguments{
\item{rel}{Reliability of the growth curve model. Either 0.8 or 0.5.}

\item{scale}{Scaling factor for the data. Default is 1, but for the growth
curve model data, it is found that having a smaller scale (1/10) can be
beneficial for convergence.}

\item{meanstructure}{Logical (this is almost always \code{FALSE}). If \code{TRUE},
include the mean structure in the estimation. Only for two-factor models
(since growth models do not have a mean structure).}

\item{x}{A data frame generated by \code{\link[=gen_data_growth]{gen_data_growth()}} or
\code{\link[=gen_data_twofac]{gen_data_twofac()}}.}
}
\value{
A named numeric vector with the true values of the model.
}
\description{
Retrieve the true values of the models used for simulation
}
\examples{
truth_growth(0.8)
truth_twofac(0.5)
}
