% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-gen_data.R
\name{truth}
\alias{truth}
\alias{truth_growth}
\alias{truth_twofac}
\title{Retrieve the true values of the models used for simulation}
\usage{
truth_growth(rel)

truth_twofac(rel, meanstructure = FALSE)
}
\arguments{
\item{rel}{Reliability of the growth curve model. Either 0.8 or 0.5.}

\item{meanstructure}{Logical. If TRUE, include the mean structure in the
estimation. Only for two-factor models (since growth models do not have a
mean structure).}
}
\value{
A named numeric vector with the true values of the model.
}
\description{
Retrieve the true values of the models used for simulation
}
\examples{
truth_growth(0.8)
truth_twofac(0.5)
}
