% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-gen_data.R
\name{two-fac}
\alias{two-fac}
\alias{txt_mod_twofac_pop}
\alias{txt_mod_twofac}
\title{Model specification for two-factor SEM models}
\usage{
txt_mod_twofac_pop(rel)

txt_mod_twofac(rel)
}
\arguments{
\item{rel}{Reliability of the growth curve model. Either 0.8 or 0.5.}
}
\value{
A character string with the model specification, used for either
fitting lavaan models or simulating data.
}
\description{
Model specification for two-factor SEM models
}
