% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-sem_rbm_functions.R
\name{fit_sem}
\alias{fit_sem}
\title{Fit a structural equation model using bias-reducing methods}
\usage{
fit_sem(
  model,
  data,
  rbm = c("implicit", "explicit", "none"),
  plugin_pen = NULL,
  debug = FALSE,
  lavfun = "sem",
  maxgrad = FALSE,
  fn.scale = 1,
  ...
)
}
\arguments{
\item{model}{A description of the user-specified model. Typically, the model
is described using the lavaan model syntax. See \link[lavaan:model.syntax]{lavaan::model.syntax} for more
information. Alternatively, a parameter table (eg. the output of the
\code{\link[lavaan:model.syntax]{lavaan::lavaanify()}} function) is also accepted.}

\item{data}{An optional data frame containing the observed variables used in
the model. If some variables are declared as ordered factors, lavaan will
treat them as ordinal variables.}

\item{rbm}{The type of RBM method to use. One of \code{"none"}, \code{"explicit"}, or
\code{"implicit"} (although, short forms are accepted, e.g. \code{"exp"}, \verb{"iRBM},
etc.)}

\item{plugin_pen}{The type of penalty to use. One of \code{NULL}, \code{"pen_ridge"},
or \code{"pen_ridge_bound"}. User-specified penalty functions are possible. See
description for details.}

\item{debug}{If TRUE, the function will return a list of intermediate results
for debugging purposes.}

\item{lavfun}{The lavaan function to use. Default is "sem".}

\item{maxgrad}{If TRUE, the function will return the maximum gradient value.}

\item{fn.scale}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} A scaling factor for the
log-likelihood function.}

\item{...}{Additional arguments to pass to the \link[lavaan:lavaan]{lavaan::lavaan} function.}
}
\value{
An object of class \code{brlavaan} which is a subclass of the
\link[lavaan:lavaan-class]{lavaan::lavaan} class.
}
\description{
This is the workhorse function to fit a structural equation model using
bias-reducing methods which we use for our simulations. For endusers, we
recommend using the \code{\link[=brsem]{brsem()}}, \code{\link[=brcfa]{brcfa()}}, or \code{\link[=brgrowth]{brgrowth()}} functions which is
a wrapper around this function and outputs something more user-friendly.
}
