% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-gen_data.R
\name{gen-data}
\alias{gen-data}
\alias{gen_data_growth}
\alias{gen_data_twofac}
\title{Generate data for simulation studies}
\usage{
gen_data_growth(
  n = 100,
  rel = 0.8,
  dist = "Normal",
  lavsim = FALSE,
  scale = 1,
  seed
)

gen_data_twofac(
  n = 100,
  rel = 0.8,
  dist = "Normal",
  lavsim = FALSE,
  meanstructure = FALSE,
  scale = 1,
  seed
)
}
\arguments{
\item{n}{Sample size}

\item{rel}{Reliability of the growth curve model. Either 0.8 or 0.5.}

\item{dist}{Distribution of the manifest variables. Either "Normal",
"Kurtosis" or "Non-normal".}

\item{lavsim}{Logical. If TRUE, simulate data using lavaan's \link[lavaan:simulateData]{lavaan::simulateData}
function.}

\item{scale}{Scaling factor for the data. Default is 1, but for the growth
curve model data, it is found that having a smaller scale (1/10) can be
beneficial for convergence.}

\item{seed}{Seed for reproducibility.}

\item{meanstructure}{Logical (this is almost always \code{FALSE}). If \code{TRUE},
include the mean structure in the estimation. Only for two-factor models
(since growth models do not have a mean structure).}
}
\value{
A data frame with simulated data. For convenience, both the true
values (\code{"truth"}) and distribution (\code{"dist"}) are attached as attributes.
}
\description{
Generate data for simulation studies.
}
\details{
The function \code{truth()} is a simple extractor function to retrieve the true
values of the data set generated by \code{gen_data_growth()} or
\code{gen_data_twofac()}.
}
\examples{
dat <- gen_data_growth(n = 100, rel = 0.8)
head(dat)
truth(dat)
}
\references{
Dhaene, S., & Rosseel, Y. (2022). Resampling Based Bias
Correction for Small Sample SEM. \emph{Structural Equation Modeling: A
Multidisciplinary Journal, 29}(5), 755–771.
https://doi.org/10.1080/10705511.2022.2057999
}
\author{
Original code by Sara Dhaene and Yves Rosseel.
}
